[##] Refacto : Moved whitespace remover to common package
